pyproject.toml
[tool.poetry]
name = "messengerbot"
version = "0.1.0"
description = ""
authors = ["JuliManhupli <juliamangup@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
